+++ library/rtsp/src/main/java/com/google/android/exoplayer2/source/rtsp/RtspMediaPeriod.java
@@ -55,6 +55,7 @@ import java.io.IOException;
 import java.net.BindException;
 import java.util.ArrayList;
 import java.util.List;
+import javax.net.SocketFactory;
 import org.checkerframework.checker.nullness.compatqual.NullableType;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 
@@ -102,6 +103,8 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
    * @param uri The RTSP playback {@link Uri}.
    * @param listener A {@link Listener} to receive session information updates.
    * @param userAgent The user agent.
+   * @param socketFactory A socket factory for {@link RtspClient}'s connection.
+   * @param debugLoggingEnabled Whether to log RTSP messages.
    */
   public RtspMediaPeriod(
       Allocator allocator,
@@ -109,6 +112,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
       Uri uri,
       Listener listener,
       String userAgent,
+      SocketFactory socketFactory,
       boolean debugLoggingEnabled) {
     this.allocator = allocator;
     this.rtpDataChannelFactory = rtpDataChannelFactory;
@@ -122,6 +126,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
             /* playbackEventListener= */ internalListener,
             /* userAgent= */ userAgent,
             /* uri= */ uri,
+            socketFactory,
             debugLoggingEnabled);
     rtspLoaderWrappers = new ArrayList<>();
     selectedLoadInfos = new ArrayList<>();